requirements.txt
# Predictive Router Oracle - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0

# Machine Learning
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Web3 / Blockchain
web3>=6.11.0

# Configuration
pyyaml>=6.0.1

# Templating
jinja2>=3.1.2

# Authentication
python-jose[cryptography]>=3.3.0

# Utilities
python-multipart>=0.0.6
pydantic>=2.5.0

# WebSocket support (included in uvicorn[standard])
websockets>=12.0

# Optional: LLM providers
openai>=1.3.0
anthropic>=0.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0